+++ src/main/java/mekanism/common/tile/multiblock/TileEntityThermalEvaporationBlock.java
@@ -1,5 +1,11 @@
 package mekanism.common.tile.multiblock;
 
+import mekanism.api.IContentsListener;
+import mekanism.api.heat.IHeatCapacitor;
+import mekanism.common.capabilities.heat.CachedAmbientTemperature;
+import mekanism.common.capabilities.holder.container.IContainerHolder;
+import mekanism.common.component.containers.type.ContainerType;
+import mekanism.common.component.containers.type.IContainerType;
 import mekanism.common.content.evaporation.EvaporationMultiblockData;
 import mekanism.common.lib.multiblock.MekanismMultiblocks;
 import mekanism.common.lib.multiblock.MultiblockType;
@@ -7,8 +13,10 @@ import mekanism.common.registries.MekanismBlocks;
 import mekanism.common.tile.prefab.TileEntityMultiblock;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
+import org.jspecify.annotations.Nullable;
 
 public class TileEntityThermalEvaporationBlock extends TileEntityMultiblock<EvaporationMultiblockData> {
 
@@ -33,5 +41,29 @@ public class TileEntityThermalEvaporationBlock extends TileEntityMultiblock<Evap
     @Override
     public boolean canBeMaster() {
         return false;
+    }
+
+    @Override
+    public boolean persists(IContainerType<?, ?> type) {
+        //But that we do not handle fluid when it comes to syncing it/saving this tile to disk
+        if (type == ContainerType.FLUID || type == ContainerType.HEAT) {
+            return false;
+        }
+        return super.persists(type);
+    }
+
+    @Nullable
+    @Override
+    protected IContainerHolder<IHeatCapacitor> getInitialHeatCapacitors(IContentsListener listener, CachedAmbientTemperature ambientTemperature) {
+        return side -> getMultiblock().getHeatCapacitors(side);
+    }
+
+    @Override
+    protected boolean onUpdateServer(ServerLevel level, EvaporationMultiblockData multiblock) {
+        boolean packet = super.onUpdateServer(level, multiblock);
+        if (multiblock.isFormed()) {
+            simulateAdjacent();
+        }
+        return packet;
     }
 }
