+++ src/main/java/mekanism/common/tile/prefab/TileEntityMultiblock.java
@@ -6,6 +6,7 @@ import mekanism.api.IConfigurable;
 import mekanism.api.IContentsListener;
 import mekanism.api.SerializationConstants;
 import mekanism.api.inventory.IInventorySlot;
+import mekanism.api.heat.IHeatHandler;
 import mekanism.api.text.EnumColor;
 import mekanism.client.SparkleAnimation;
 import mekanism.common.MekanismLang;
@@ -371,6 +372,16 @@ public abstract class TileEntityMultiblock<T extends MultiblockData> extends Til
     @Override
     public InteractionResult onSneakRightClick(Level level, Player player) {
         return InteractionResult.PASS;
+    }
+
+    @Nullable
+    @Override
+    public IHeatHandler getAdjacent(Direction side) {
+        BlockPos relativePos = getBlockPos().relative(side);
+        if (getMultiblock().locations.contains(relativePos)) {
+            return null;
+        }
+        return super.getAdjacent(side);
     }
 
     //Methods relating to IComputerTile
